fifo Queue Implmentation
Implementation of the first in first out queue.
Following functions have been implmented:
add_Q
remove_Q
get_Q

queue adds data in the form of a pointer to void.
So, this header file can be included to any program and
the program will be able to use the queue to store data.
The structure of the data will be known to the calling
application, not to the header file.
header file will just add references to the next and prev
data structures in the queue, but cannot dereference the
data elements itself.